new:created a defult MainLayout.jsx file so side bar will always get included in new routes
