+++ frontend/src/MainLayout.jsx
@@ -1,0 +1,13 @@
+import { ReactNode } from 'react'
+import SidebarComponent from '@/components/Sidebar';
+
+ const MainLayout= ({children}) => {
+    return(
+        <div className='flex flex-row'>
+            <SidebarComponent/>
+            <main>{children}</main>
+        </div>
+    )
+}
+
+export default MainLayout
